Deprecate unused methods
