+++ Dnn.CommunityForums/class/ForumBase.cs
@@ -31,21 +31,14 @@ namespace DotNetNuke.Modules.ActiveForums
         #region Private Member Variables
 
         private int? _forumId;
-        private string _forumIds = string.Empty;
         private int? _forumGroupId;
-        private int _parentForumId = -1;
         private int? _postId;
         private int? _topicId; // = -1;
         private int? _replyId;
         private int? _quoteId;
         private int? _authorid;
         private bool? _jumpToLastPost;
-        private string _defaultView = Views.ForumView;
-        private int _defaultForumViewTemplateId = -1;
-        private int _defaultTopicsViewTemplateId = -1;
-        private int _defaultTopicViewTemplateId = -1;
         private string _templatePath = string.Empty;
-        private string _templateFile = string.Empty;
         private Forum _foruminfo;
         private XmlDocument _forumData;
 
@@ -73,74 +66,14 @@ namespace DotNetNuke.Modules.ActiveForums
 
         public ControlsConfig ControlConfig { get; set; }
 
-        public string ThemePath
-        {
-            get
-            {
-                return Page.ResolveUrl(MainSettings.ThemeLocation);
-            }
-        }
+        public string ThemePath => Page.ResolveUrl(MainSettings.ThemeLocation);
 
-        public string ForumIds
-        {
-            get
-            {
-                return _forumIds;
-            }
-            set
-            {
-                _forumIds = value;
-            }
-        }
+        public string ForumIds { get; set; } = string.Empty;
 
-        public int DefaultForumViewTemplateId
-        {
-            get
-            {
-                return _defaultForumViewTemplateId;
-            }
-            set
-            {
-                _defaultForumViewTemplateId = value;
-            }
-        }
-        public int DefaultTopicsViewTemplateId
-        {
-            get
-            {
-                return _defaultTopicsViewTemplateId;
-            }
-            set
-            {
-                _defaultTopicsViewTemplateId = value;
-            }
-        }
-
-        public int DefaultTopicViewTemplateId
-        {
-            get
-            {
-                return _defaultTopicViewTemplateId;
-            }
-            set
-            {
-                _defaultTopicViewTemplateId = value;
-            }
-        }
-
-        public string DefaultView
-        {
-            get
-            {
-
-                return _defaultView;
-            }
-            set
-            {
-
-                _defaultView = value;
-            }
-        }
+        public int DefaultForumViewTemplateId { get; set; } = -1;
+        public int DefaultTopicsViewTemplateId { get; set; } = -1;
+        public int DefaultTopicViewTemplateId { get; set; } = -1;
+        public string DefaultView { get; set; } = Views.ForumView;
 
         public bool JumpToLastPost
         {
@@ -422,29 +355,9 @@ namespace DotNetNuke.Modules.ActiveForums
             }
         }
 
-        public int ParentForumId
-        {
-            get
-            {
-                return _parentForumId;
-            }
-            set
-            {
-                _parentForumId = value;
-            }
-        }
+        public int ParentForumId { get; set; } = -1;
 
-        public string TemplateFile
-        {
-            get
-            {
-                return _templateFile;
-            }
-            set
-            {
-                _templateFile = value;
-            }
-        }
+        public string TemplateFile { get; set; } = string.Empty;
 
         public Forum ForumInfo
         {
@@ -459,29 +372,8 @@ namespace DotNetNuke.Modules.ActiveForums
         }
 
         public int SocialGroupId { get; set; }
-
-        public bool CanRead
-        {
-            get
-            {
-                if(!_canRead.HasValue)
-                    _canRead =  SecurityCheck("read");
-
-                return _canRead.Value;
-            }
-        }
-
-        public bool CanView
-        {
-            get
-            {
-                if(!_canView.HasValue)
-                    _canView = SecurityCheck("view");
-
-                return _canView.Value;
-            }
-        }
-
+        public bool CanRead => _canRead ?? SecurityCheck("read");
+        public bool CanView => _canView ?? SecurityCheck("view");
         public bool CanCreate
         {
             get
